+++ models/schemas/cars.js
@@ -28,7 +28,14 @@ const model = db.MAIN_DB.model(
         imgUrls: [{ type: String }], // * Array of Images 
         createTime: { type: Date, default: Date.now() }, // ? Time this car dataset was created - [ defaults to current date if not provided ]
         updateTime: { type: Date, default: Date.now() }, // ? Time this car dataset was updated - [ defaults to current date if not provided ]
-        bookedDates: { type: Array, required: false } // * Dates when the car is already booked by users
+        bookedDates: { type: Array, required: false }, // * Dates when the car is already booked by users
+        orders: [
+            { 
+                type: mongoose.Schema.Types.ObjectId, 
+                ref: 'order', 
+                required: false 
+            }
+        ] // * all orders of this car
     })
 )
 module.exports = model;