+++ release-notes.md
@@ -142,3 +142,4 @@
     - Added a `force=1` parameter which will create a new Invoice even if one exists.
 - Added Cancellation Policy page.
 - Order Bookings now reflect on usersDB
+- Booking page now dynamically takes booked dates from the ordersDB by using the carID as foriegn key query parameter.